+++ tests/RequestBuilder/EventsRequestBuilderTest.php
@@ -51,6 +51,8 @@ class EventsRequestBuilderTest extends TestCase
         $builder->addUserMerge($userMergeCommand1);
         $builder->addUserMerges([$userMergeCommand2, $userMergeCommand3]);
 
+        $builder->setRequestId('custom-request-id-foo');
+
         $request = $builder->build();
 
         $this->assertInstanceOf(Request::class, $request);
@@ -68,6 +70,8 @@ class EventsRequestBuilderTest extends TestCase
         $this->assertContains($userMergeCommand1, $requestData);
         $this->assertContains($userMergeCommand2, $requestData);
         $this->assertContains($userMergeCommand3, $requestData);
+
+        $this->assertSame('custom-request-id-foo', $request->getRequestId());
     }
 
     /** @test */